Changed project structure
Project structure changed to be more modular and expandable.
Eearly organization reflected a testing stage where it was my first experience with the api.
All source code is put in the "src" folder and executable binaries in the "bin" folder.
Packages will be stored in the "src" folder as well.